WIP - IDEMPOTENT
Idempotency issues were uncovered upstream, there is a lot of work to be
done on the individual pivotal cookbooks. All the `package` resources
break when installed twice, and there are some duplicated pieces in the
recipes. Not too bad, but still, I think we're going to need to work
through those steps since they are the ones that caused me to add all
the rewind stuff in sprout-homebrew-emacs -- this setup appears to work
though, with the sprout-homebrew-emacs.